+++ cq/ConstantQ.h
@@ -60,7 +60,7 @@ public:
     virtual int getLatency() const { return m_outputLatency; } 
     virtual double getMaxFrequency() const { return m_p.maxFrequency; }
     virtual double getMinFrequency() const;
-    virtual double getBinFrequency(int bin) const;
+    virtual double getBinFrequency(double bin) const; // bin may be nonintegral
 
     /**
      * Given a series of time-domain samples, return a series of
